test(integration): migrate to new multi-store API
