+++ tests/integration/src/tests/playwright/fixtures/devtools/todomvc/components/Footer.tsx
@@ -1,10 +1,9 @@
 /** biome-ignore-all lint/a11y: testing */
 import { queryDb } from '@livestore/livestore'
-import { useStore } from '@livestore/react'
 import type React from 'react'
-
 import { uiState$ } from '../livestore/queries.ts'
 import { events, tables } from '../livestore/schema.ts'
+import { useAppStore } from '../livestore/store.ts'
 import type { Filter } from '../types.ts'
 
 const incompleteCount$ = queryDb(tables.todos.count().where({ completed: false, deletedAt: null }), {
@@ -12,7 +11,7 @@ const incompleteCount$ = queryDb(tables.todos.count().where({ completed: false, 
 })
 
 export const Footer: React.FC = () => {
-  const { store } = useStore()
+  const store = useAppStore()
   const { filter } = store.useQuery(uiState$)
   const incompleteCount = store.useQuery(incompleteCount$)
   const setFilter = (filter: Filter) => store.commit(events.uiStateSet({ filter }))
